rename salsa variables and files to mailchimp
