+++ mailchimp_auth/constants.py
