+++ mailchimp_auth/models.py
