--- salsa_auth/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class SalsaAuthConfig(AppConfig):
-    name = 'salsa_auth'
